Initial add of new entrypoint key variant
